refactored file name to be coherent with class name
