+++ pennylane_snowflurry/__init__.py
@@ -1,2 +1,2 @@
-from .snowflurry_converter import PennylaneConverter
+from .pennylane_converter import PennylaneConverter
 from .snowflurry_device import SnowflurryQubitDevice
